Small overall changes and some fixes.
Blocks now overlap by one sample.
Bug in 'delete montage' fixed.
Memory size option changed.
Now cache gets cleared after filter change.


Former-commit-id: 0fbdfcd2bbe791e4951f5a515507ed20e5e3ba96
Former-commit-id: ba82f94c949abb4e2b3201e845220351d975362b

+++ App/src/DataFile/gdf2.cpp
@@ -201,10 +201,10 @@ case a_:\
 	samplingFrequency = vh.samplesPerRecord[0]/duration;
 
 	// Construct the cache.
-	unsigned int memoryAvailable = PROGRAM_OPTIONS["dataFileCacheSize"].as<unsigned int>();
+	int64_t memoryAvailable = PROGRAM_OPTIONS["dataFileCacheSize"].as<int64_t>();
 	if (memoryAvailable > 0)
 	{
-		cache = new QCache<unsigned int, std::vector<char>>(max(static_cast<unsigned int>(1), memoryAvailable/(vh.samplesPerRecord[0]*getChannelCount()*dataTypeSize)));
+		cache = new QCache<unsigned int, std::vector<char>>(max(1, static_cast<int>(memoryAvailable/(vh.samplesPerRecord[0]*getChannelCount()*dataTypeSize))));
 	}
 }
